Create main.css
This file imports all the styling for the content in the front-end pages. The first ANCHOR at the top of the file imports the 2 fonts that are being used from Google Fonts ( a free website font service). The second tag resets all the styling so the website becomes standard, then all the tags after that modify them to the styles I wanted. These can be moved.